% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{aggr_coexpr_within_dataset}
\alias{aggr_coexpr_within_dataset}
\title{Aggregate cell type correlation matrices within a single dataset}
\usage{
aggr_coexpr_within_dataset(
  mat,
  meta,
  pc_df,
  cor_method,
  agg_method,
  min_cell = 20,
  verbose = TRUE
)
}
\arguments{
\item{mat}{A sparse gene by cell matrix}

\item{meta}{A data frame that maps cell IDs to cell types}

\item{pc_df}{A data frame of unique protein coding gene symbols}

\item{cor_method}{One of "pearson" or "spearman"}

\item{agg_method}{One of "allrank", "colrank", or "FZ"}

\item{min_cell}{A non-negative integer corresponding to how many cells of a
cell type must have at least one count for a gene to be considered "measured"
and thus viable for calculating correlation}

\item{verbose}{Logical determining if the current cell type and time should
be displayed as a message during iteration}
}
\value{
A list of two gene by gene matrices, the first of which is the
aggregate coexpression matrix and the second of which is the NA tracking
matrix
}
\description{
Iterates through each unique cell type in meta, subsetting mat to cells of
the given type and generating a cell type coexpression matrix for each,
which are aggregated into a single matrix.
}
\details{
https://pubmed.ncbi.nlm.nih.gov/27165153/
https://pubmed.ncbi.nlm.nih.gov/25717192/
https://pubmed.ncbi.nlm.nih.gov/34015329/

all_rank: each cell type coexpression matrix is ranked across the entire
matrix, then summed and rank standardized into \link{0, 1}.

col_rank: each cell type coexpression matrix is ranked column-wise, then
summed and rank standardized into \link{0, 1}.

FZ: Fisher's Z transformation is applied to each cell type coexpression
matrix, which are then summed and divided element-wise by the count of times
a given gene-gene pair was co-measured across all cell types.
}
