% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{diag_to_one}
\alias{diag_to_one}
\title{Set diag of a square matrix to 1}
\usage{
diag_to_one(mat)
}
\arguments{
\item{mat}{A dense numeric matrix}
}
\value{
A matrix of the same dimensions where all values are preserved save
for the diagonals assigned to 1
}
\description{
Set diag of a square matrix to 1
}
