% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{fisherz}
\alias{fisherz}
\title{Fisher Z transformation}
\usage{
fisherz(cor)
}
\arguments{
\item{cor}{A numeric correlation value (single element, array, matrix)}
}
\value{
The transformed correlations
}
\description{
Fisher Z transformation
}
\seealso{
DescTools::FisherZ()
}
