% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{load_scdat}
\alias{load_scdat}
\title{Load a single cell dataset}
\usage{
load_scdat(path)
}
\arguments{
\item{path}{A character corresponding to the path of the dataset}
}
\value{
A list of the single cell count matrix and cell type metadata
}
\description{
Assumes path leads to an RDS of a list with 2 named elements: "Mat", the
sparse count matrix, and "Meta", a metadata data frame mapping cell IDs to an
annotated cell type
}
