% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{lowertri_to_symm}
\alias{lowertri_to_symm}
\title{Replace the upper triangle of square matrix with its lower triangle.}
\usage{
lowertri_to_symm(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A matrix of the same dimensions where all upper triangular values are
replaced by its lower triangle values.
}
\description{
Replace the upper triangle of square matrix with its lower triangle.
}
