% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Set NAs in matrix to 0}
\usage{
na_to_zero(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A matrix of the same dimensions where all values are preserved save
for NAs assigned to 0
}
\description{
Set NAs in matrix to 0
}
