% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{sparse_pcor}
\alias{sparse_pcor}
\title{Column-wise Pearson's correlation for sparse matrices}
\usage{
sparse_pcor(mat)
}
\arguments{
\item{mat}{A sparse numeric m by n Matrix}
}
\value{
A dense n by n matrix of the Pearson's correlation between the
columns of mat
}
\description{
Column-wise Pearson's correlation for sparse matrices
}
