% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{transform_correlation_mat}
\alias{transform_correlation_mat}
\title{Transform correlation matrix}
\usage{
transform_correlation_mat(cmat, agg_method)
}
\arguments{
\item{cmat}{A dense gene by gene correlation matrix}

\item{agg_method}{One of "allrank", "colrank", or "FZ"}
}
\value{
A dense gene by gene matrix of transformed correlations
}
\description{
Take an input correlation matrix and transform it to prepare for aggregation.
Transformation entails all ranking, column ranking, or Fisher's Z.
Each methods impute NA cors to 0 and ensure that the diagonal equals 1.
}
\details{
allrank: make the matrix tri. to prevent double ranking symmetric elements
and then jointly rank standardize the matrix into \link{0,1} such that higher pos.
correlations are closer to 1.

colrank: rank standardize each column separately into \link{0,1} such that higher
pos. correlations are closer to 1.

FZ: perform Fisher's Z transform on the raw correlations
}
