% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{uppertri_to_na}
\alias{uppertri_to_na}
\title{Set upper triangle of a square matrix to NA}
\usage{
uppertri_to_na(mat)
}
\arguments{
\item{mat}{A matrix}
}
\value{
A matrix of the same dimensions where all values are preserved save
for the values above the diagonal set to NA and the diagonal is preserved
}
\description{
Set upper triangle of a square matrix to NA
}
