% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{prepare_celltype_mat}
\alias{prepare_celltype_mat}
\title{Prepare a cell type matrix for coexpression}
\usage{
prepare_celltype_mat(mat, meta, pc_df, cell_type, min_cell = 20)
}
\arguments{
\item{mat}{A sparse gene by cell matrix}

\item{meta}{A data frame that maps cell IDs to cell types}

\item{pc_df}{A data frame of unique protein coding gene symbols}

\item{cell_type}{A character of the cell type to subset mat on}

\item{min_cell}{A non-negative integer corresponding to how many cells of a
cell type must have at least one count for a gene to be considered "measured"}
}
\value{
A sparse cell by gene matrix
}
\description{
Subset mat to cells annotated to the input cell_type, set sparse genes to 0
and transpose the resulting matrix.
}
