% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_functions.R
\name{transform_correlation_mat}
\alias{transform_correlation_mat}
\title{Transform correlation matrix}
\usage{
transform_correlation_mat(cmat, agg_method)
}
\arguments{
\item{cmat}{A gene by gene correlation matrix}

\item{agg_method}{One of "allrank", "colrank", or "FZ"}
}
\value{
A gene by gene matrix of transformed correlations
}
\description{
Take an input correlation matrix and transform it to prepare for aggregation.
Transformation entails either all ranking, column ranking, or Fisher's Z.
Each methods impute NA cors to 0 and ensure that the diagonal equals 1.
}
\details{
allrank: make the matrix tri. to prevent double ranking symmetric elements
and then jointly rank the matrix (lower rank = positive cor).

colrank: rank each column separately (lower rank = positive cor).

FZ: perform Fisher's Z transform on the raw correlations
}
